move image size and num channels extraction to functional
