+++ torchvision/prototype/transforms/_auto_augment.py
@@ -9,7 +9,7 @@ from torchvision.prototype.transforms import Transform, InterpolationMode, AutoA
 from torchvision.prototype.utils._internal import apply_recursively
 from torchvision.transforms import functional as _F
 
-from ._utils import query_image, get_image_size, get_image_num_channels
+from ._utils import query_image
 
 
 @dataclasses.dataclass
@@ -108,7 +108,7 @@ class _AutoAugmentBase(Transform):
         fill = self.fill
         if isinstance(fill, (int, float)):
             image = query_image(sample)
-            num_channels = get_image_num_channels(image)
+            num_channels = F.get_image_num_channels(image)
             fill = [float(fill)] * num_channels
         elif fill is not None:
             fill = [float(f) for f in fill]
@@ -123,7 +123,7 @@ class _AutoAugmentBase(Transform):
         params = params or self._get_params(sample)
 
         image = query_image(sample)
-        image_size = get_image_size(image)
+        image_size = F.get_image_size(image)
 
         for transform_id, magnitude in self.get_transforms_meta(image_size):
             dispatcher = self._DISPATCHER_MAP[transform_id]
