+++ torchvision/prototype/transforms/functional/__init__.py
@@ -1,4 +1,5 @@
 from torchvision.transforms import InterpolationMode  # usort: skip
+from ._utils import get_image_size, get_image_num_channels  # usort: skip
 from ._meta_conversion import convert_bounding_box_format, convert_color_space  # usort: skip
 
 from ._augment import (
